Moved regrid bilinear array unit tests.
